bug-fix: Push Project2 and update project urls
When project directories were renamed, links to the specific projects in GitHub were changed.
This will resolve the link issue by pointing links to correct urls.

Previous commit ignored project 2 as there was Wrangling in the .gitignore file.
This line was deleted and project 2 should be added to the repository